+++ igor-plugin-message/src/main/java/com/arassec/igor/plugin/message/connector/EmailBaseConnector.java
@@ -19,47 +19,47 @@ public abstract class EmailBaseConnector extends BaseMessageConnector {
      * Hostname of the host providing an E-Mail service (e.g. IMAP or SMTP).
      */
     @NotEmpty
-    @IgorParam(value = 1)
+    @IgorParam(sortIndex = 1)
     protected String host;
 
     /**
      * Username for authentication.
      */
     @NotEmpty
-    @IgorParam(value = 3)
+    @IgorParam(sortIndex = 3)
     protected String username;
 
     /**
      * Password for authentication.
      */
     @NotEmpty
-    @IgorParam(value = 4, secured = true)
+    @IgorParam(sortIndex = 4, secured = true)
     protected String password;
 
     /**
      * Set to {@code true} to enable TLS.
      */
-    @IgorParam(value = 5, advanced = true)
+    @IgorParam(sortIndex = 5, advanced = true)
     protected boolean enableTls = true;
 
     /**
      * If {@code true}, always accepts the provided SSL/TLS certificate from the server.
      */
-    @IgorParam(value = 6, advanced = true)
+    @IgorParam(sortIndex = 6, advanced = true)
     protected boolean alwaysTrustSsl;
 
     /**
      * Connection timeout in milliseconds.
      */
     @Positive
-    @IgorParam(value = 7, advanced = true)
+    @IgorParam(sortIndex = 7, advanced = true)
     protected int connectionTimeout = 5000;
 
     /**
      * Read timeout in milliseconds.
      */
     @Positive
-    @IgorParam(value = 8, advanced = true)
+    @IgorParam(sortIndex = 8, advanced = true)
     protected int readTimeout = 15000;
 
     /**